pyproject.toml
[project]
name = "ops-assistant"
version = "0.1.0"
description = "Bounded ops assistant for Datadog triage"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "google-genai>=1.0.0",
    "langgraph>=0.2.0",
    "langgraph-checkpoint>=2.0.0",
    "ddtrace>=3.1.0",
    "datadog>=0.50.0",
    "datadog-api-client>=2.30.0",
    "fastmcp>=2.13.0",
    "structlog>=24.0.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "ragas==0.1.21",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.35.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ops_triage_agent", "ops_triage_mcp_server", "sas_generator", "sas_mcp_server", "dashboard_mcp_server", "dashboard_enhancer", "shared"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.ruff.lint.per-file-ignores]
"ops_triage_agent/prompts/*.py" = ["E501"]  # Allow long lines in prompt templates
"sas_generator/prompts.py" = ["E501"]  # Allow long lines in prompt templates
"sas_mcp_server/data/*.py" = ["E501"]  # Allow long lines in data definitions
"dashboard_enhancer/designer/gemini_designer.py" = ["E501"]  # Allow long lines in prompts

[dependency-groups]
dev = [
    "playwright>=1.57.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.10",
]
